### 2018-07-09 #816 @whedon
[PRE REVIEW]: YAD: A Learning-based Inductive Logic Programming Tool
**Submitting author:** @hamedmx (<a href="http://orcid.org/0000-0002-9132-593X">Hamed Karimi</a>)
**Repository:** <a href="https://github.com/hamedmx/YAD-ILP-Tool" target ="_blank">https://github.com/hamedmx/YAD-ILP-Tool</a>
**Version:** v1.0
