### 2018-07-09 374f503
Update paper.md
