### 2018-07-09 7e53518
Update paper.md
