### 2018-07-09 7ebbc7a
Update paper.md
