### 2018-07-09 8d11b5d
Update paper.md
