+++ paper.md
@@ -78,9 +78,9 @@ The architecture of this ILP tool is shown in Figure 2. As it is obvious in this
  **User Interface:** It provides some necessary facilities and UI in order to present the input data and configure the model parameters for user. These user-configurable parameters were mentioned before in previous sections.<br/>
  **Rule Preparation:** This section has some tasks to read the input file, parse the information, and produce background knowledge and positive and negative examples set in appropriate form of the first-order logic format. All of the examples (positive and negative examples) are located in the format of a set of objects. Also, the division of the input data into the train and test sets is accomplished by this module.<br/>
  **Induction Engine:** This section forms the main part of the tool, which consists of three subsections:
- - *Generalizer:* This subsection has a task to produce more general logical rules from positive examples. In the other words, this module receives a positive example and tries to generalize it as much as possible.
- - *Rule Checker:* The logical rules that are generated by generalizer, are controlled by this module to determine whether they satisfy the problem conditions (covering the positive examples and not covering the negative examples) or not. So, if the conditions of the desired logical rule are not covered, then it will not be accepted.
- - *Pruning Module:* This module has a task to prevent from the over-fitting of the model, otherwise, we probably have many logical rules that each of them satisfies only one or two positive examples, i.e. they do not have an adequate generality. Pruning operates in such a way that not only we can achieve the best results in terms of precision, but also the number of logical rules will be minimized.<br/>
+   - *Generalizer:* This subsection has a task to produce more general logical rules from positive examples. In the other words, this module receives a positive example and tries to generalize it as much as possible.
+   - *Rule Checker:* The logical rules that are generated by generalizer, are controlled by this module to determine whether they satisfy the problem conditions (covering the positive examples and not covering the negative examples) or not. So, if the conditions of the desired logical rule are not covered, then it will not be accepted.
+   - *Pruning Module:* This module has a task to prevent from the over-fitting of the model, otherwise, we probably have many logical rules that each of them satisfies only one or two positive examples, i.e. they do not have an adequate generality. Pruning operates in such a way that not only we can achieve the best results in terms of precision, but also the number of logical rules will be minimized.<br/>
  **Evaluation Module:** It has a task to execute the logical rules produced by induction module on the test data and evaluates the results. For this reason, the various measurements such as Precision, Recall, Accuracy, and F-Measure are computed and reported.<br/>
  **Shared Data:** The sections we have mentioned before, have access to a set of data which is the input and output data of ILP algorithm such as background knowledge set, positive and negative examples, and the set of produced logical rules by the algorithm.
 
