### 2018-07-09 9ab9fa5
Update paper.md
