+++ paper.md
@@ -26,16 +26,16 @@ bibliography: paper.bib
 Inductive logic learning is considered as one of the most prominent approaches for multi-dimensional and multi-tabular learning by relying on first-order logic to describe train data in the form of positive and negative examples which can find a set of existing rules and relations and can propose them in first-order logic format [@Muggleton1995], [@Lavrac1994].
 Generally, the inductive learning problem is defined as follows:
  - **Input:**
-    + Background knowledge, B, a set of Horn clauses.
-    + Positive examples E^+, a set of Horn clauses (typically ground literals).
-    + Negative examples E^-, a set of Horn clauses (typically ground literals).
+     + Background knowledge, B, a set of Horn clauses.
+     + Positive examples E^+, a set of Horn clauses (typically ground literals).
+     + Negative examples E^-, a set of Horn clauses (typically ground literals).
  - **Output:** 
-    + A hypothesis, H, a set of Horn clauses.
+     + A hypothesis, H, a set of Horn clauses.
  - **Problem Conditions:** 
-    + All the positive examples should be producible from combination of background knowledge and H, which it indicates the completeness criterion:<br/>
+     + All the positive examples should be producible from combination of background knowledge and H, which it indicates the completeness criterion:<br/>
     
 			$ \forall e \in E^+\colon B \wedge H \longrightarrow e $    (Completeness)<br/>
-    + None of the negative examples should be producible, which it indicates the consistency criterion and also obviously keeps the Precision measure high:<br/>
+     + None of the negative examples should be producible, which it indicates the consistency criterion and also obviously keeps the Precision measure high:<br/>
     
 			$$*\forall e \in E^-\colon B \wedge H \longrightarrow e*$$    (Consistency)<br/>
 	
@@ -53,10 +53,10 @@ Generally, this algorithm (Algorithm 1) has been presented as follows:
 ![](https://github.com/hamedmx/YAD-ILP-Tool/blob/master/algo%201.PNG)
 
 This tool has been implemented in C# and a screenshot of its GUI1 is depicted in Figure 1. The input arguments of YAD are:
- - The percentage of Train Examples (the remaining percentage will be Test Examples percentage for testing the generated generic logical rules on the data-set).
- - The number of steps (usually, it is set to the default value 4) which is used to generate a general predicate in every step of induce function (inverse resolution operations) in order to produce generic logical rules. So, for example, with four steps, the tool can produce general logical rules with at most four predicates in their right-hand side.
- - Try Count (usually, it is set to the default value 5) is the number of endeavors that the tool performs to create and generate a general logical rule.
- - Negative Threshold (usually, it is set to a low amount e.g. the default value 5) is the percentage of the whole negative examples that every produced generic logical rule can produce negative examples equal or lower than the selected percentage, otherwise, the produced logical rule is useless and it is not considered as an output generic logical rule.
+- The percentage of Train Examples (the remaining percentage will be Test Examples percentage for testing the generated generic logical rules on the data-set).
+- The number of steps (usually, it is set to the default value 4) which is used to generate a general predicate in every step of induce function (inverse resolution operations) in order to produce generic logical rules. So, for example, with four steps, the tool can produce general logical rules with at most four predicates in their right-hand side.
+- Try Count (usually, it is set to the default value 5) is the number of endeavors that the tool performs to create and generate a general logical rule.
+- Negative Threshold (usually, it is set to a low amount e.g. the default value 5) is the percentage of the whole negative examples that every produced generic logical rule can produce negative examples equal or lower than the selected percentage, otherwise, the produced logical rule is useless and it is not considered as an output generic logical rule.
 Also, the use cases and functions of this tool are as follows (respectively in use):
[…]
