### 2018-07-09 fd066cd
Update paper.md
