### 2018-07-09 f19d38b
Update paper.md
