+++ paper.md
@@ -25,17 +25,17 @@ bibliography: paper.bib
 
 Inductive logic learning is considered as one of the most prominent approaches for multi-dimensional and multi-tabular learning by relying on first-order logic to describe train data in the form of positive and negative examples which can find a set of existing rules and relations and can propose them in first-order logic format [@Muggleton1995], [@Lavrac1994].
 Generally, the inductive learning problem is defined as follows:
- - **Input:**
-     + Background knowledge, B, a set of Horn clauses.
-     + Positive examples E^+, a set of Horn clauses (typically ground literals).
-     + Negative examples E^-, a set of Horn clauses (typically ground literals).
- - **Output:** 
-     + A hypothesis, H, a set of Horn clauses.
- - **Problem Conditions:** 
-     + All the positive examples should be producible from combination of background knowledge and H, which it indicates the completeness criterion:<br/>
+- **Input:**
+    + Background knowledge, B, a set of Horn clauses.
+    + Positive examples E^+, a set of Horn clauses (typically ground literals).
+    + Negative examples E^-, a set of Horn clauses (typically ground literals).
+- **Output:** 
+    + A hypothesis, H, a set of Horn clauses.
+- **Problem Conditions:** 
+    + All the positive examples should be producible from combination of background knowledge and H, which it indicates the completeness criterion:
     
 			$ \forall e \in E^+\colon B \wedge H \longrightarrow e $    (Completeness)<br/>
-     + None of the negative examples should be producible, which it indicates the consistency criterion and also obviously keeps the Precision measure high:<br/>
+    + None of the negative examples should be producible, which it indicates the consistency criterion and also obviously keeps the Precision measure high:
     
 			$$*\forall e \in E^-\colon B \wedge H \longrightarrow e*$$    (Consistency)<br/>
 	
@@ -57,6 +57,7 @@ This tool has been implemented in C# and a screenshot of its GUI1 is depicted in
 - The number of steps (usually, it is set to the default value 4) which is used to generate a general predicate in every step of induce function (inverse resolution operations) in order to produce generic logical rules. So, for example, with four steps, the tool can produce general logical rules with at most four predicates in their right-hand side.
 - Try Count (usually, it is set to the default value 5) is the number of endeavors that the tool performs to create and generate a general logical rule.
 - Negative Threshold (usually, it is set to a low amount e.g. the default value 5) is the percentage of the whole negative examples that every produced generic logical rule can produce negative examples equal or lower than the selected percentage, otherwise, the produced logical rule is useless and it is not considered as an output generic logical rule.<br/>
+
 Also, the use cases and functions of this tool are as follows (respectively in use):
 1. **Open:** Opening a file (usually, a text file) consists of Background Knowledge (the lines should start with ‘B’), Positive Examples (the lines should start with ‘+’), and Negative Examples (the lines should start with ‘-’) in first-order logic format as logical rules.
 2. **Induce:** Producing general logical rules by inducing (using inverse resolution) on background knowledge.
@@ -81,6 +82,7 @@ The architecture of this ILP tool is shown in Figure 2. As it is obvious in this
  - *Generalizer:* This subsection has a task to produce more general logical rules from positive examples. In the other words, this module receives a positive example and tries to generalize it as much as possible.
  - *Rule Checker:* The logical rules that are generated by generalizer, are controlled by this module to determine whether they satisfy the problem conditions (covering the positive examples and not covering the negative examples) or not. So, if the conditions of the desired logical rule are not covered, then it will not be accepted.
[…]
