### 2018-07-09 5b995ad
Update paper.md
