### 2018-07-10 51c4a7b
Update paper.md
