+++ paper.md
@@ -26,19 +26,19 @@ bibliography: paper.bib
 Inductive logic learning is considered as one of the most prominent approaches for multi-dimensional and multi-tabular learning by relying on first-order logic to describe train data in the form of positive and negative examples which can find a set of existing rules and relations and can propose them in first-order logic format [@Muggleton1995], [@Lavrac1994].
 Generally, the inductive learning problem is defined as follows:
 - **Input:**
-    + Background knowledge, B, a set of Horn clauses.
+    + Background knowledge, *B*, a set of Horn clauses.
     + Positive examples $E^+$, a set of Horn clauses (typically ground literals).
     + Negative examples $E^-$, a set of Horn clauses (typically ground literals).
 - **Output:** 
-    + A hypothesis, H, a set of Horn clauses.
+    + A hypothesis, *H*, a set of Horn clauses.
 - **Problem Conditions:**
-    + All the positive examples should be producible from combination of background knowledge and H, which it indicates the completeness criterion:
+    + All the positive examples should be producible from combination of background knowledge and *H*, which it indicates the completeness criterion:
 
-			*$\forall e \in E^+\colon B \wedge H \to e$*    (Completeness)
+			$\forall e \in E^+\colon B \wedge H \to e$    (Completeness)
     + None of the negative examples should be producible, which it indicates the consistency criterion and also obviously keeps the Precision measure high:
 
-			*$\forall e \in E^-\colon B \wedge H \to e$*    (Consistency)
-It is remarkable to say that the symbols $E^+$, $E^-$, B, and H are proposed in first-order logic format.<br/>
+			$\forall e \in E^-\colon B \wedge H \to e$    (Consistency)
+It is remarkable to say that the symbols $E^+$, $E^-$, *B*, and H are proposed in first-order logic format.<br/>
 Due to the expressiveness and high flexibility of first-order logic, the inductive learning approach has high capability in discovering the relations. Also, since the set of logic’s rules is not restricted to a specific table or structure, it can be used in widespread areas and problems like finding relationships in social networks and discovering communication patterns in protein structures (bioinformatics problems).<br/>
 Despite of all the advantages that learning based on inductive logic has; in practice, it becomes a problem to search and find the optimal solution which with respect to the vastness of search space, it has no polynomial and deterministic solution. So, in order to solve it, we need to the heuristic and randomized algorithms.<br/>
 In recent years, many researches were performed in this area and numerous algorithms and tools were proposed such as FOIL [@Quinlan1993], [@Quinlan1990], Golem, ProGolem [@Muggleton2010], Aleph [@Srinivasan2007], and CIGOL.
@@ -50,6 +50,7 @@ Generally, this algorithm (Algorithm 1) has been presented as follows:
 > Algorithm 1. The proposed induction algorithm used in YAD ILP tool
 
 ![](https://github.com/hamedmx/YAD-ILP-Tool/blob/master/algo%201.PNG)
+> Algorithm 1. The proposed induction algorithm used in YAD ILP tool
 
 This tool has been implemented in C# and a screenshot of its GUI1 is depicted in Figure 1. The input arguments of YAD are:
 - The percentage of Train Examples (the remaining percentage will be Test Examples percentage for testing the generated generic logical rules on the data-set).
@@ -78,9 +79,9 @@ The architecture of this ILP tool is shown in Figure 2. As it is obvious in this
  **User Interface:** It provides some necessary facilities and UI in order to present the input data and configure the model parameters for user. These user-configurable parameters were mentioned before in previous sections.<br/>
[…]
