### 2018-07-10 98a85ff
Update paper.md
