### 2018-07-10 f1ffc0f
Update paper.md
