### 2018-07-10 a0f36b9
Update paper.md
