### 2018-07-10 f28138b
Update paper.md
