+++ paper.md
@@ -50,7 +50,7 @@ Generally, this algorithm (Algorithm 1) has been presented as follows:
 
 > **Algorithm 1.** The proposed induction algorithm used in YAD ILP tool
 ![](https://github.com/hamedmx/YAD-ILP-Tool/blob/master/algorithm%201.PNG)
-> *Constructing a generalized logical rule using inverse resolution on a predicate P with background knowledge B and a hash map h.*
+> *\*Constructing a generalized logical rule using inverse resolution on a predicate P with background knowledge B and a hash map h.*
 
 This tool has been implemented in C# and a screenshot of its GUI1 is depicted in Figure 1. The input arguments of YAD are:
 - The percentage of Train Examples (the remaining percentage will be Test Examples percentage for testing the generated generic logical rules on the data-set).
