### 2018-07-10 67ae750
Update paper.md
