### 2018-07-10 1295b37
Update paper.md
