### 2018-07-10 5266cf0
Update paper.md
