### 2018-07-10 b6d1ce3
Update paper.md
