### 2018-07-10 #816 @hamedmx
Hello @arfon and @jasonclark , first of all thank you for your guidance. well, it seems to have some problems to compile my paper.md file i guess and i could not find them to resolve. please help me in compile process.
