### 2018-07-10 dd2c498
Update paper.md
