### 2018-07-10 f568e06
Update paper.md
