+++ paper.md
@@ -22,10 +22,10 @@ bibliography: paper.bib
 
 Inductive logic learning is considered as one of the most prominent approaches for multi-dimensional and multi-tabular learning by relying on first-order logic to describe train data in the form of positive and negative examples which can find a set of existing rules and relations and can propose them in first-order logic format [@Muggleton1995], [@Lavrac1994].
 Generally, the inductive learning problem is defined as follows:
-- **Input:**
-    + Background knowledge, *B*, a set of Horn clauses.
-    + Positive examples $E^+$, a set of Horn clauses (typically ground literals).
-    + Negative examples $E^-$, a set of Horn clauses (typically ground literals).
+* **Input:**
+    * Background knowledge, *B*, a set of Horn clauses.
+    * Positive examples $E^+$, a set of Horn clauses (typically ground literals).
+    * Negative examples $E^-$, a set of Horn clauses (typically ground literals).
 - **Output:** 
     + A hypothesis, *H*, a set of Horn clauses.
 - **Problem Conditions:**
@@ -71,8 +71,10 @@ The architecture of this ILP tool is shown in Figure 2. As it is obvious in this
 ![](https://github.com/hamedmx/YAD-ILP-Tool/raw/master/fig%202.PNG)
 > **Figure 2.** ``YAD`` ILP tool architecture
 
- **User Interface:** It provides some necessary facilities and UI in order to present the input data and configure the model parameters for user. These user-configurable parameters were mentioned before in previous sections.<br/>
- **Rule Preparation:** This section has some tasks to read the input file, parse the information, and produce background knowledge and positive and negative examples set in appropriate form of the first-order logic format. All of the examples (positive and negative examples) are located in the format of a set of objects. Also, the division of the input data into the train and test sets is accomplished by this module.<br/>
+ **User Interface:** It provides some necessary facilities and UI in order to present the input data and configure the model parameters for user. These user-configurable parameters were mentioned before in previous sections.
+ 
+**Rule Preparation:** This section has some tasks to read the input file, parse the information, and produce background knowledge and positive and negative examples set in appropriate form of the first-order logic format. All of the examples (positive and negative examples) are located in the format of a set of objects. Also, the division of the input data into the train and test sets is accomplished by this module.<br/>
+
  **Induction Engine:** This section forms the main part of the tool, which consists of three subsections:
  - ***Generalizer:*** This subsection has a task to produce more general logical rules from positive examples. In the other words, this module receives a positive example and tries to generalize it as much as possible.
  - ***Rule Checker:*** The logical rules that are generated by generalizer, are controlled by this module to determine whether they satisfy the problem conditions (covering the positive examples and not covering the negative examples) or not. So, if the conditions of the desired logical rule are not covered, then it will not be accepted.
