### 2018-07-10 #816 @arfon
> Thank you so much @arfon . Now, i have some problems; firstly, in writing mathematical statements and their symbols and secondly, in going to a new line, using bullets, and writing captions for the figures. So, is there any links to the related guidelines ?

We're just using standard Pandoc/Markdown here: https://pandoc.org/README.pdf

Looks like the document is compiling now :-)
