### 2018-07-10 f82003b
Update paper.md
