### 2018-07-10 8a05706
Update paper.md
