### 2018-07-10 0f7424e
Update paper.md
