### 2018-07-10 b8e7300
Update paper.md
