### 2018-07-10 11f4b8c
Update paper.md
