+++ paper.bib
@@ -43,7 +43,6 @@ year = {2018}
 @techreport{Srinivasan2007,
 author = {Srinivasan, Ashwin},
 title = {{The Aleph Manual}},
-url = {https://scholar.google.com/citations?user=zpNTeogAAAAJ{\&}hl=en{\&}oi=sra{\#}d=gs{\_}md{\_}cita-d{\&}p={\&}u={\%}2Fcitations{\%}3Fview{\_}op{\%}3Dview{\_}citation{\%}26hl{\%}3Den{\%}26user{\%}3DzpNTeogAAAAJ{\%}26citation{\_}for{\_}view{\%}3DzpNTeogAAAAJ{\%}3Ad1gkVwhDpl0C{\%}26tzom{\%}3D-270},
 year = {2007}
 }
 @incollection{Quinlan1993,
