### 2018-07-10 b5c789e
Update paper.md
