### 2018-07-10 b5fe1ed
Update paper.md
