### 2018-07-10 4f9e75d
Update paper.md
