### 2018-07-10 dd0ee0d
Update paper.md
