### 2018-07-10 b6757cf
Update paper.md
