### 2018-07-10 2f9010c
Update paper.md
