### 2018-07-20 #816 @jasonclark
@hamedmx Can you create a README file for the repository? All JOSS entries require a general description of the software (what is it?, why would I use it?, etc.) along with basic install instructions. You might even use some of the narrative from your article at https://github.com/hamedmx/YAD-ILP-Tool/blob/master/YAD-JOSS.pdf.
