### 2018-07-21 5e6697c
Update paper.md
