### 2018-07-21 #816 @hamedmx
@jasonclark The required README file is ready in the repository with adequate descriptions about the tool and its usage and also a new additional sample test file has been uploaded there.
