### 2018-08-05 #816 @arfon
> @arfon , @jasonclark has declared his willing to edit this submission since about 16 days ago and according to his revision on our repository, we have already prepared the desired README file for the repository. So, when should we identify the reviewers for our submission and start the actual review ?

Yes, time to identify reviewers. @hamedmx - could you take a look a this list of potential reviewers and suggest a few people you think might be suitable? https://bit.ly/joss-reviewers
