### 2018-08-09 #816 @martinmodrak
I've so far never reviewed for JOSS, so I am not sure if I am expected to react to @hamedmx 's suggestion or if I should wait for @jasonclark . Anyway, I feel the software is on the border of my knowledge (honestly, probably some distance on the wrong side of the border). Also I will be out of office for most of August and early September and so will be able to have a first look no sooner than ~ September 7th. Just thought this might be factors in reviewer selection.
