### 2018-08-10 #816 @arfon
> @martinmodrak Thank you very much for your notification. @arfon So, what is the next step to begin the actual review section ? I've already suggested some other reviewers for our submission about 5 days ago.

Hi @hamedmx, we just need a little more time here I think. @martinmodrak isn't able to look at this submission until early September and none of the other suggested reviewers have responded yet so I think we're still in the pre-review phase.

@xirdneh  @fusionet24 @dredamonsta1 - would one of you be willing to review this submission for JOSS?
