### 2018-08-13 #816 @dredamonsta1
Sorry I’m late to reply my mailbox was full I can start reviewing later today I am also new and not sure but I want to help out as best as I can
