### 2018-08-13 #892 @whedon
[REVIEW]: YAD: A Learning-based Inductive Logic Programming Tool
**Submitting author:** @hamedmx (<a href="http://orcid.org/0000-0002-9132-593X">Hamed Karimi</a>)
**Repository:** <a href="https://github.com/hamedmx/YAD-ILP-Tool" target ="_blank">https://github.com/hamedmx/YAD-ILP-Tool</a>
**Version:** v1.0
**Archive:** <a href="https://doi.org/10.6084/m9.figshare.7189043" target="_blank">10.6084/m9.figshare.7189043</a>
