### 2018-08-13 #816 @jasonclark
Closing out of pre-review and setting review at https://github.com/openjournals/joss-reviews/issues/892
