### 2018-08-13 #892 @jasonclark
@dredamonsta1 the YAD software is ready for your review.
cc @arfon @hamedmx
