### 2018-08-13 #892 @jasonclark
@dredamonsta1 the review checklist is above, but let me know if you have questions.
