### 2018-08-20 #892 @dredamonsta1
everything is in tact as ive seen
