### 2018-08-20 #892 @hamedmx
@dredamonsta1 Thank you. @arfon Is this issue completed by our reviewer's response ? What is the next step ?
