### 2018-08-22 #892 @hamedmx
@arfon Thank you for reopening this review issue. As a concern, Is this issue in the actual review process by @dredamonsta1 for JOSS now or it is stopped ?
