### 2018-08-30 #892 @cmccomb
@jasonclark I have a lot of upcoming travel. I could spend some time on this after 9/24, but I won't be able to turn around a good review before then. Please advise!
