### 2018-09-26 #892 @cmccomb
@jasonclark, please advise!
