### 2018-10-01 #892 @jasonclark
@cmccomb Thanks, Chris! @arfon You mentioned that we might need to reopen. It looks like that has happened. Can I just reassign Chris as editor suing whedon? cc @hamedmx
