### 2018-10-01 #892 @arfon
> @cmccomb Thanks, Chris! @arfon You mentioned that we might need to reopen. It looks like that has happened. Can I just reassign Chris as editor using whedon? cc @hamedmx

Kind of. I'll need to handle this semi-manually.
