### 2018-10-05 hamedmx/YAD-ILP-Tool#1 @cmccomb
Community Guidelines
Per the reviewer checklist, there should be clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support. I couldn't find these in the README file. If they're somewhere else, it would be really helpful to have them at the top level!
