### 2018-10-05 #892 @cmccomb
@hamedmx I think there's only one thing that needs to be fixed: adding some guidelines for potential contributors, bug-reporting, etc. (see Documentation > Community Guidelines in the reviewer checklist above). With that, I think that my review will be complete.
https://github.com/hamedmx/YAD-ILP-Tool/issues/1
