### 2018-10-05 #892 @hamedmx
@cmccomb Thank you very much.
So, can I add those community guidelines to the README.md file or, I should add it to the original paper.md file?
