### 2018-10-07 #892 @hamedmx
@cmccomb I provided the desired community guidelines in README.md file
