### 2018-10-07 #892 @cmccomb
@hamedmx Thanks for making those changes!
@jasonclark This completes my review!
