### 2018-10-10 #892 @hamedmx
@arfon As I can see the information for authors in author guidelines of JOSS, I created an account in figshare to upload the updated submission and got a DOI for this submission. What should I do now?
