### 2018-10-10 #892 @arfon
@jasonclark - I think this is good to accept now. Can you confirm?
